renamed most of the files to use kebab-case and correct index files to set some convention
