+++ src/components/schedule-manager/add-course-modal.vue
