+++ src/components/schedule-manager/course-list-item.vue
