+++ src/components/scroll-bar-hider.vue
