+++ src/model/exam-timerange.js
