+++ src/model/schedule-header.js
