+++ src/model/schedule-item.js
